Move validation helper functions to validation.hpp #1458
