+++ libraries/protocol/include/steemit/protocol/steem_operations.hpp
@@ -2,22 +2,11 @@
 #include <steemit/protocol/base.hpp>
 #include <steemit/protocol/block_header.hpp>
 #include <steemit/protocol/asset.hpp>
+#include <steemit/protocol/validation.hpp>
 
-#include <fc/utf8.hpp>
 #include <fc/crypto/equihash.hpp>
 
 namespace steemit { namespace protocol {
-
-   inline void validate_account_name( const string& name )
-   {
-      FC_ASSERT( is_valid_account_name( name ), "Account name ${n} is invalid", ("n", name) );
-   }
-
-   inline void validate_permlink( const string& permlink )
-   {
-      FC_ASSERT( permlink.size() < STEEMIT_MAX_PERMLINK_LENGTH, "permlink is too long" );
-      FC_ASSERT( fc::is_utf8( permlink ), "permlink not formatted in UTF8" );
-   }
 
    struct account_create_operation : public base_operation
    {
